pyproject.toml
[project]
name = "repl-nix-legaldataassistant"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.9.0",
    "groq>=0.11.0",
    "nltk>=3.9.1",
    "openai>=1.52.2",
    "plotly>=5.24.1",
    "scikit-learn>=1.5.2",
    "seaborn>=0.13.2",
    "spacy>=3.8.2",
    "streamlit>=1.39.0",
    "wordcloud>=1.9.3",
]
